Initial - 12/29/2024
Creates a movement path as a json to be interpreted by an auto program. This compiler is intended for creating drivetrain movement paths and minimal arm adjustments. This compiler will not interact with servos or timed delays.

This compiler creates short, reusable jsons representing paths between a few key points. This should not be used to create an entire auto program all in one part.

Movement paths are created using Catmull-Rom splines, generating smoother paths. The more key steps used, the more control over the path of the robot. First passes are NOT expected to produce viable results, iteration will be necessary.

Queer Coded by Striker-909. If you use this class or a method from this class in its entirety, please make sure to give credit.